Split data.scala in two.
